support filetype tags
contains a dictionary to map form vim filetype identifiers,
to the "scope" names Sublime uses in its Syntax files.
This was generated half-automatically, and scope names
cover those available in Sublime and the CMake plugin.

The mapping from Sublime's "scope" names to package names
if created the first time it is requested.

both mappings should allow changes by the user
(not implemented).

Signed-off-by: Norbert Lange <nolange79@gmail.com>